zip: Allow backslash as path separator
Some Windows zip archivers, such as PowerShell's Compress-Archive, use
backslash as the path separator in violation of the zip specification.
Info-Zip works around this by supporting this for archives that report
"version made by" of 0 (MS-DOS).

This change adds the same workaround to libarchive and adds a test of
this case using a zip file created with PowerShell.