Make mtree less eager to bid
At least one keyword is required to consider a file an mtree archive.
Moreover the search will stop if no keyword is found in the first 6 lines.
This prevents misidentification of a simple CSV files as mtree archives.
CSV file shall be identified as "raw" - test has been added.
The original use-case was processing of gzipped CSV files with libarchive.
The decompression went fine but then the mtree format bid for it ... and failed.