Check if root before operations.
No overall change to the code path when root; when not root, ensure
created files are initially user writable if we need to do XATTRs or HFS
compression. Mode is fixed later to the desired final mode in this case.

Net of one extra syscall per file to fix the mode at the end only when
all of these are true:
 * Not root
 * Final mode is not owner-writable
 * XATTRs or HFS compression are needed.

These changes make it unexpectedly pass the xattr test on freebsd.