RAR5 reader: fix invalid formatting strings ('%02x' -> '%x')
archive_set_error() uses its own implementation of sprintf-like format
specificators, and one of the specificators that is not supported is
padded numeric value (i.e. %02x). The result is that when an error
occurs, libarchive returns a string that contains an unprocessed format
string instead of processed one, like in this example:

    Error: Unsupported filter type: 0x%02x

where it should be:

    Error: Unsupported filter type: 0x1F

This commit changes padded numeric format specificators into simple
un-padded numeric specificators, for which libarchive's own sprintf
implementation has support for.