RAR5 reader: window_mask was not updated correctly
The `window_mask` variable should be always in sync with the
`window_size` variable.

The commit fixes a bug in which there was one place that `window_size`
was modified, but `window_mask` wasn't updated. This was leading to a
SIGSEGV error, because by having wrong `window_mask`, RAR5 reader was
accessing the memory outsize current window buffer.

The commit also adds a test for this issue, together with OSSFuzz
sample #15278.