Introduce ARCHIVE_EXTRACT_ATOMIC and set it by default on bsdtar.
This flag changes the way that regular files are extracted:

Instead of removing existing files first and re-creating them in
order to replace their contents, a temporary file is created and
when writing to the temporary file is completed, the file is
rename(2)d to the final destination name.

This has the effect of presenting a consistent view of the file to
the system (either the file with the new contents or the file with
the old contents). Removing and overwriting the file has the
undesired side effect that the the system can either not see the
file at all (from the time it is being removed till the time it is
being re-created), or worse it can see partial file contents. This
is problematic when extracting system files (for example shared
libraries).

There are two issues with this patch (there could be more, but these
are the two that come to my mind):

1. mkostemp(3) is not portable. One approach is to provide our own
   that also does the right chmod. The other is a mess of ifdefs
   depending on what flavor of mk{s,}temp(2) the system provides. It
   is also questionable what to do on non-unix systems.
2. I set it by default on bsdtar, perhaps there should be a flag
   to disable it, when for example it is not desirable because of
   space constraints. I am not sure what to call it.